elasticsearch_connector/requirements.txt
certifi==2022.6.15
elastic-transport==8.1.2
elasticsearch==8.3.1
kafka-python==2.0.2
urllib3==1.26.10
